# swagger/auth.yaml
openapi: 3.0.3
info:
  title: Dompet.in Auth API
  version: 1.0.0
  description: Endpoints for user registration, login, and token check.
servers:
  - url: http://localhost:3000/api
paths:
  /register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation or duplicate email error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      tags: [Auth]
      summary: Login user and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      tags: [Auth]
      summary: Get current user info by JWT
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, nickname, password]
      properties:
        email:
          type: string
          format: email
          example: farhan@mail.com
        nickname:
          type: string
          example: Farhan
        password:
          type: string
          minLength: 6
          example: secret123
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
        token:
          type: string
          description: Returned if auto-login is enabled
        user:
          $ref: '#/components/schemas/User'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: farhan@mail.com
        password:
          type: string
          example: secret123
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    MeResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id_user:
          type: integer
          example: 1
        email:
          type: string
          example: farhan@mail.com
        nickname:
          type: string
          example: Farhan
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid email or password